#12. Timer for game was added. It's all. Some 'feature' will be  added in this game, but later...

+++ app/src/main/java/com/example/duska/axelerom/ResultsActivity.java
@@ -42,7 +42,7 @@ public class ResultsActivity extends AppCompatActivity {
                 //используем индекс для получения строки
                 String currentPlayerName = cursor.getString(playerNameColumnIndex);
                 arrayNames.add(currentPlayerName);
-                double currentTime = cursor.getInt(timeColumnIndex);
+                double currentTime = cursor.getDouble(timeColumnIndex);
                 arrayTime.add(currentTime);
                 //выведем в логи таблицу
                 Log.d("mLog", "name = " + arrayNames +
